refactor:🌈 Test용 프로토콜 위치 변경
- HistoryListSupport, TravelListSupport 파일을 생성하여 테스트를 위한 프로토콜,
  혹은 enum 등과 같은 데이터 구조 코드를 옮겨놓음
